# libs/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TSDHN API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Health",
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/version": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Version",
        "operationId": "version_api_v1_version_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calculations": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Calculation",
        "operationId": "create_calculation_api_v1_calculations_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EarthquakeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job",
        "operationId": "create_job_api_v1_jobs_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{app_job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_v1_jobs__app_job_id__get",
        "parameters": [
          {
            "name": "app_job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{app_job_id}/events": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Job Events",
        "operationId": "job_events_api_v1_jobs__app_job_id__events_get",
        "parameters": [
          {
            "name": "app_job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{app_job_id}/report": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Report",
        "operationId": "get_job_report_api_v1_jobs__app_job_id__report_get",
        "parameters": [
          {
            "name": "app_job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "307": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CalculationPreview": {
        "properties": {
          "calculation": {
            "$ref": "#/components/schemas/CalculationResponse"
          },
          "travel_times": {
            "$ref": "#/components/schemas/TsunamiTravelResponse"
          }
        },
        "type": "object",
        "required": [
          "calculation",
          "travel_times"
        ],
        "title": "CalculationPreview",
        "description": "Preview data returned before committing to a queued simulation."
      },
      "CalculationResponse": {
        "properties": {
          "length": {
            "type": "number",
            "title": "Length"
          },
          "width": {
            "type": "number",
            "title": "Width"
          },
          "dislocation": {
            "type": "number",
            "title": "Dislocation"
          },
          "seismic_moment": {
            "type": "number",
            "title": "Seismic Moment"
          },
          "tsunami_warning": {
            "type": "string",
            "title": "Tsunami Warning"
          },
          "distance_to_coast": {
            "type": "number",
            "title": "Distance To Coast"
          },
          "azimuth": {
            "type": "number",
            "title": "Azimuth"
          },
          "dip": {
            "type": "number",
            "title": "Dip"
          },
          "epicenter_location": {
            "type": "string",
            "title": "Epicenter Location"
          },
          "rectangle_parameters": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Rectangle Parameters"
          },
          "rectangle_corners": {
            "items": {
              "additionalProperties": {
                "type": "number"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Rectangle Corners"
          }
        },
        "type": "object",
        "required": [
          "length",
          "width",
          "dislocation",
          "seismic_moment",
          "tsunami_warning",
          "distance_to_coast",
          "azimuth",
          "dip",
          "epicenter_location",
          "rectangle_parameters",
          "rectangle_corners"
        ],
        "title": "CalculationResponse"
      },
      "EarthquakeInput": {
        "properties": {
          "Mw": {
            "type": "number",
            "title": "Mw"
          },
          "h": {
            "type": "number",
            "title": "H"
          },
          "lat0": {
            "type": "number",
            "title": "Lat0"
          },
          "lon0": {
            "type": "number",
            "title": "Lon0"
          },
          "dia": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dia",
            "default": "00"
          },
          "hhmm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hhmm",
            "default": "0000"
          }
        },
        "type": "object",
        "required": [
          "Mw",
          "h",
          "lat0",
          "lon0"
        ],
        "title": "EarthquakeInput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthStatus": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "database_connected": {
            "type": "boolean",
            "title": "Database Connected"
          },
          "storage_connected": {
            "type": "boolean",
            "title": "Storage Connected"
          }
        },
        "type": "object",
        "required": [
          "status",
          "timestamp",
          "database_connected",
          "storage_connected"
        ],
        "title": "HealthStatus"
      },
      "JobCreated": {
        "properties": {
          "app_job_id": {
            "type": "string",
            "title": "App Job Id"
          },
          "compute_job_id": {
            "type": "string",
            "title": "Compute Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "result_bucket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Bucket"
          },
          "result_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Key"
          }
        },
        "type": "object",
        "required": [
          "app_job_id",
          "compute_job_id",
          "status"
        ],
        "title": "JobCreated"
      },
      "JobRequest": {
        "properties": {
          "app_job_id": {
            "type": "string",
            "format": "uuid",
            "title": "App Job Id"
          },
          "input": {
            "$ref": "#/components/schemas/EarthquakeInput"
          },
          "skip_steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skip Steps"
          }
        },
        "type": "object",
        "required": [
          "app_job_id",
          "input"
        ],
        "title": "JobRequest"
      },
      "JobStatusResponse": {
        "properties": {
          "app_job_id": {
            "type": "string",
            "title": "App Job Id"
          },
          "compute_job_id": {
            "type": "string",
            "title": "Compute Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step"
          },
          "step_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Index"
          },
          "total_steps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Steps"
          },
          "calculation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CalculationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "travel_times": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TsunamiTravelResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "result_bucket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Bucket"
          },
          "result_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Key"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "report_available": {
            "type": "boolean",
            "title": "Report Available",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "app_job_id",
          "compute_job_id",
          "status"
        ],
        "title": "JobStatusResponse"
      },
      "TsunamiTravelResponse": {
        "properties": {
          "arrival_times": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Arrival Times"
          },
          "distances": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Distances"
          },
          "epicenter_info": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Epicenter Info"
          }
        },
        "type": "object",
        "required": [
          "arrival_times",
          "distances",
          "epicenter_info"
        ],
        "title": "TsunamiTravelResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "name",
          "version"
        ],
        "title": "VersionInfo"
      }
    }
  }
}